Fixed floating point errors in calculator LUT. Minmax now attempts to evaluate card playing before active power as it probably makes more sense (not sure in a real game). It still evaluates it, but it's more likely for a lethal play to involve playing cards rather than active power, so it's a pruning help probably. Also Added test to verify EOT is immediate if there's no play options, and a test that finds lethal in hypothetical draws succesfully

+++ OdlEngine/GameInstance/CalculatorLut.cs
@@ -34,7 +34,7 @@ namespace GameInstance
             {
                 for (int i = 0; i < handSize; i++) // Draw a sample for each card in hand, calculate the chance of NOT drawing
                 {
-                    result *= (deckSize - cardCount - i) / (deckSize - i); // Each draw, deck shrinks
+                    result *= (float)(deckSize - cardCount - i) / (deckSize - i); // Each draw, deck shrinks
                 }
                 result = 1 - result; // Calculate the chance of drawing
             }
@@ -57,7 +57,7 @@ namespace GameInstance
         {
             int deckSize = parameters.Item1;
             int cardCount = parameters.Item2;
-            return cardCount / deckSize;
+            return (float)cardCount / deckSize;
         }
         /// <summary>
         /// Square root of integer.
